4. A server being "less blocking" by using threads
This server uses a blocking socket to interact with clients like in example 1,
but will pass handling of the request and response to a thread so that the main
thread is free to immediately accept another connection. accept() is still blocking,
but there is much less blocking now that most of the processing per client is in its own thread.